backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
PyPDF2==3.0.1
google-generativeai==0.7.2
faiss-cpu==1.9.0.post1
numpy==2.1.0
python-dotenv==1.0.1
tiktoken==0.8.0